# openapi/Penguin.yaml
openapi: 3.0.3
info:
  version: 0.6.0
  title: Penguin API
  description: "A list of the currently provided Interfaces to interact with the penguin, implementing file operations and sending messages"
security:
  - {}
...
servers:
  - url: "http://{apiRoot}:{port}/v1"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local pen node main API
      port:
        default: "1633"
        description: Service port provided in pen node config
  - url: "http://{apiRoot}:{port}"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local pen node main API
      port:
        default: "1633"
        description: Service port provided in pen node config
paths:
...
  "/pen":
    post:
      summary: "Upload file or a collection of files"
      description: "In order to upload a collection, user can send a multipart request with all the files populated in the form data with appropriate headers.\n\n
        User can also upload a tar file along with the penguin-collection header. This will upload the tar file after extracting the entire directory structure.\n\n
        If the penguin-collection header is absent, all requests (including tar files) are considered as single file uploads.\n\n
        A multipart request is treated as a collection regardless of whether the penguin-collection header is present. This means in order to serve single files
        uploaded as a multipart request, the penguin-index-document header should be used with the name of the file."
      tags:
        - File
        - Collection
      parameters:
        - in: query
          name: name
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/FileName"
          required: false
          description: Filename when uploading single file
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinTagParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPinParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinEncryptParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/ContentTypePreserved"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinCollection"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinIndexDocumentParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinErrorDocumentParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPostageBatchId"
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
          application/x-tar:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Ok
          headers:
            "penguin-tag":
              $ref: "PenguinCommon.yaml#/components/headers/PenguinTag"
            "etag":
              $ref: "PenguinCommon.yaml#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "403":
          $ref: "PenguinCommon.yaml#/components/responses/403"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/pen/{reference}":
    patch:
      summary: "Reupload a root hash to the network"
      tags:
        - Reupload file
        - Reupload collection
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/PenguinReference"
          required: true
          description: Root hash of content
      responses:
        "200":
          description: Ok
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
    get:
      summary: "Get file or index document from a collection of files"
      tags:
        - Collection
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/PenguinReference"
          required: true
          description: Penguin address of content
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinRecoveryTargetsParameter"
      responses:
        "200":
          description: Ok
          headers:
            "penguin-recovery-targets":
              $ref: "PenguinCommon.yaml#/components/headers/PenguinRecoveryTargets"
          # "penguin-feed-index":
          #   $ref: "PenguinCommon.yaml#/components/headers/PenguinFeedIndex"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "404":
          $ref: "PenguinCommon.yaml#/components/responses/404"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/pen/{reference}/{path}":
    get:
      summary: "Get referenced file from a collection of files"
      tags:
        - Collection
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/PenguinReference"
          required: true
          description: Penguin address of content
        - in: path
          name: path
          schema:
            type: string
          required: true
          description: Path to the file in the collection.
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinRecoveryTargetsParameter"
      responses:
        "200":
          description: Ok
          headers:
            "penguin-recovery-targets":
              $ref: "PenguinCommon.yaml#/components/headers/PenguinRecoveryTargets"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "404":
          $ref: "PenguinCommon.yaml#/components/responses/404"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
...